Introduce a game context
Why:
We will like to eventually share game state among many components. A
react Context is a good way to do this.

This commit:
Initializes a game context with a simple counter for the application.

This game might be complicated enough to justify use of redux, though it
would probably be best to see how far we can get with just context
objects before making the switch to something heavier.