Implement basic piece color
Why:
We would like to be able to distinguish between black and white pieces.

This commit:
Add the logic to initialize a piece with a color and render that colors
piece. Currently we are just using some applied styles to color a view
component, it would likely be better to render an image asset for the
pieces in a future commit.