Select and move pieces
Why:
We would like users to be able to move their pieces around on the board.

This commit:
Adds the logic in the game context for users selecting tiles and moving
pieces conditionally based off of the user tile selection. Also
introduced was the naming convention of rowIdx and colIdx to make the
references to the piece coordinates more consistent throughout the app.
RowIdx and ColIdx refer to the index of the ultimate component rendered,
either the BoardRow, or the tile within the row respectively.