Conditionally render piece color
Why:
We would like for the white and black pieces to render with the white
and black versions of their image assets

This commit:
introduces some logic in the Piece Component to check the piece color
before choosing which component to render. Note, it might make sense to
abstract this logic into a dedicated Pawn Component.

Co-authored-by: Dan Huber <dsh1204@gmail.com>