Computer selects from own pieces
Why:
Currently the computer is just selecting tiles from random on the board.
We would like for the computer to select from its own pieces.

This commit:
Adds the logic for the computer to select randomly from its own pieces
and then select a random valid move for that piece. Note if there is no
valid move for that piece the computer will simply unselect its
selection.