Introduce pieces module and knights
Why:
We would like to introduce the knight piece into the game.

This commit refactors the pieces constructors to all be in the same file.
Also we went back to using react native styles to differentiate the
pieces graphically. Importing the assets proved to be a bit cumbersome
to implement and was slowing down the development. We will come back to
adding assets in a later commit.