Add bishop piece
We would like for users to be able to play with a bishop piece.

Also reafactored was the logic for creating the initial board.

Note that the bishops movement currently allows for jumping over
blocking pieces. A future commit will need to handle the logic for not
considering moves that jump blocking pieces valid for a bishop.