Remove expo update rn version
Why:
Currently we are behind a few versions of react-native and we'd like to
be on the most recent, 0.60. Updating expo proved cumbersome and we know
we'd like to use native code so it makes sense to just remove expo
altogether at this point in addition to updating the rn version

This commit:
Removes expo, updates rn to 0.60, and also update react-navigation to
the most recent version.