Use chess.js to do move validations
Why:
We would like to levarge the third party library chess.js to handle move
validation for the chess game, rather than writing our own bespoke chess
engine which could be time intensive, prone to bugs, and maintainance
burdens.

This commit:

- Adds the chess.js library

- Creates some intermediate interface methods to interact with the
chess.js api. Specifically the chess.js engine expects boards to be
represented in Forsyth-Edwards Notation and moves to be presented in
Alegebraic Chess Notation (e.g. "e4") so we needed to build some
serialization helpers for this purpose.

- Replaces the current logic in GameHelpers.validMove to use the
chess.js engine.

Note that this has also made it possible for the computer ai to select a
piece that has no valid moves, and leads to a situation where the ai
does nothing during the game. Fixing this will come in a future pr.