Separate piece style from logic and improve types
Why:
Currenlty the piece objects are responsible for both game logic and
stylistic concerns. This is a point of coupling which will make it
challenging to do refactors and development of the game logic without
affecting the styles and vice versa. We would like to keep these
concerns separate.

This commit:
Removes the styles from the piece objects, moves them into a
`styles/pieces.ts` file, and introduces logic to the rn component for
correctly rendering the right style for the pieces. The piece type
information was also improved a bit as well.