remove back button and restyle board screen
Why:
For the MVP the user won't have a need to return to the welcome screen,
so removing that button reduces the screen clutter. Styling the page to
match the approved designs.

This commit:
Removes the go back button and it's functionality, adds an image and new
styling. Note: the status bar still needs to be inverted, will cut a new
card for that.