Computer ai selects valid moves at random
Why:
Currently the ai can get into a state were it selects a piece to move
that has no valid moves and it stops playing. We would like for the ai
to select from all valid moves for all pieces at random to avoid this
issue.

This commit:
Updates the GameHelpers.validMoves function to use the chess.js engine
to provide the game context all of the possible valid moves for all of
the pieces. The GameContext was updated to select from these Moves. Note
that we are still using the selectTile mechanic for the ai to mimic the
feel of playing against a human that selects and deselects tiles to
move.