Players can win or lose
Why:
We would like for players to be able to win if they are able to take
their opponents king. And they should lose if they lose their king.

This commit:
Adds a new function in the game helpers called winner which will iterate
over the board and determine if any player has lost their king yet and
declare a winner if so. This function is used in the game context tick
function to see if the game is over.

@testing-library/jest-native was added to use some better matchers in
the spec for determine if the game over message had been rendered.

Note that currently the chess.js chess engine can only generate valid
moves with what it considers a valid game state and it will throw errors
with some game states. For example, it can error if you provide it a
game state where its a players turn and that player does not have a
king. Checking if the king is present before delegating to the chess.js
engine solves that particular issue, but it is still possible to throw
runtime errors with chess.js