add game over screen
Why:
When the player wins or loses the game, they will need to see a clear
indication that the game is over and be prompted to start a new game.

This commit:

Adds a GameOver component that the GameScreen will render if there is a
winner present in the game context. It adds images to represent the
victory state, and refactors the primary button to be more reusable.

Co-Authored-By johnschoeman <johnschoeman1617@gmail.com>