Refactor piece style application
Why:
In a future commit we would like to switch out the current method of
building pieces images by applying view styles with image assets.

This commit:
Refactors the Piece component such that it will be easier to make this
change. No functionality has been added or changed.

Also caught in this pr was an opportunistic bug fix in the validMoves
function. there are some board configurations that will cause the
chess.js chess client to throw an error even though the client will
validate that configuration as valid.