Ai will take a piece if possible
Why:
In order for the game to be more engaging the computer ai needs to be
more challenging. Currently it's just moving pieces at random. We would
like for it to take pieces if possible.

This commit:
Introduces an ai_helpers module to move the compute ai logic into. A new
function, selectMove, was added to be used to determine the computers
next move. selectMove will find all the valid moves, determine if any
are attacking and if so, will choose an attacking move at random. If no
attacking moves are found, it will select a valid move at random.

Note that the chess.js type declarition for Move limits the capture
field to all piece types but the king type. However it is possible for a
move from chess.js to be return that has 'caputure: "k"', hence I needed
to redifine the Move interface locally.