Use pngs for piece images
Why:
Currently we are using view styles to build peice graphics. We would
like for our piece graphics to look nice and be clear on what the piece
type is.

This commit:
Adds some pngs for the piece type and updates the Piece component to
render with the piece image instead of the view styles. We also removed
the old piece image assets and font since they were no longer being
used.

Co-Authored-By: jeffstolz <jstolz123@gmail.com>