Pawn can promote to queens
Why:
Currently pawns are not allow to move to the back row. We would like for
pawns to be automatically promoted to queens when they enter the back
row.

This commit:
Updates the validMove function to consider pawns moving to the back row as
valid and updates the updateBaord function to promote pawns to queens as
they move onto the back row.