Select piece if own
Why:
Currently, if a user has a piece selected and they select on of their
own pieces, nothing will happen. This is a confusing mechanic fro users.

This commit:
Changes the game behaviour when selecting a piece. If a user has a piece
selected and their next selection is one of their own pieces, thier
piece selection will change to the new piece.

Also .metals was added to .gitignore becasue vscode.