Add types signatures to chess logic
Why:
A previous commit required us to in line the logic we were using for our
chess engine so that we could modify its behaviour easily. Since the
chess engine was plain, untyped javascript there were many type errors
caught by the ts compilier. We did not fix those errors at in the
previous commit for sake of time and to be able to ship with smaller
more focused commits.

This commit:
Removes unused code, adds type signatures and adds some refactoring to
cover previously uncaught edge cases. No additional functionality was
added in this commit.