Allow kings to move into check
Why:
Currently the game will not allow players to move their kings into check
or move make any move that doesn't bring the king out of check if the
king is in check because these are considered illegal moves in
traditional chess. For real time chess, however, it is confusing if you
cannot do this. It would be a better game experience if you could move
your king in and out of check and the game only ends if one player
manages to take the opponents king.

This commit:
Moves the logic from the chess.js engine into the code base and alerts
the generate_moves function to allow for these illegal moves.
Unfortunately the chess.js engine does not expose any api to allow for
this hence the need to fork the code.  It could have been possible to
fork the repo, but we felt it was easier to copy the code over, since it
was just one file.
Note that the chess.js code is not typed and it's a lot of work to type
it out properly so we are currently in a state where the ts compiler
will fail. We are shipping now as is for sake of time and will take on
the work of updating the code for typescript in a following commit.