Move rook if player castles king
Why:
We would like for players to be able to castle their king. Currently
players can move their king the two spaces in a valid castle move, but
the board will not correctly update with the rook passing through the
king.

This commit:
Updates the updateBoard function to handle the case of moving a rook
during a castle. The validMove function was also provided some specs
around considering castling moves valid or invalid.

Note that there is currently a bug in which players can move either
thier king or rook and still be able to castle afterwards. This is due
to the fact that the chess engine relies on game state to determine this
rule. so only the rules of no intermediate spaces being occupied or
attacked is enforced. We will need to capture the remaining rule of no
movement of the king or rook prior to castle in a later commit.