Introduce ai that evaluates board value
Why:
Currently our ai makes random moves. We would like for users to be able
to play against a computer that tries to find the highest value move.

This commit:
introduces a new chess ai, maximizing_ai, to the game and switches the
ai used in the game context. Note we are leaving the logic for the
random_ai as is since we are planning to add a feature of choosing
different computer difficulties in a future commit.