Fix ai castling bug
Why:
Currently there is a bug in the chess engine where if the computer has
already castled it's possible for the computer to return a castling move
as a valid move. For some reason when this happens, the make_pretty
method will return a move with a none value value for 'to', either "8"
or "".

This commit:
adds simple check on the length of 'to' to determine if this is a valid
move or not. This is certainly a hack and we should look for a better
solution in a future commit. This could also be solved by managing the
castling state for the players and providing that state to the provided
fen code.