Allow for multiple game types
Why:
We would like to be able to A/B test various game rules and allow
players to choose which game rules they'd like to play with.

This commit:
Introduces an ArcadeContext to handle the logic of which game type a
player is playing and refactors current winner the GameContext to
accomdate. A new set of folders where introduced within Game/ to
contain a sample option of different game types. Note that currently
both Classic and ThreeKings are identical, adding the different game
rules will come in a future commit.