Remove react navigation and refactor screen nav
Why:
Currently the app can support various game types from the previous
commit, but during that feature addition much of the screens and
layouts became broken visually.

This commit:
Fixes the broken visual aspects and removes react-navigation. React
Navigaiton is a large combursome library and we were not using the vast
majority of its features. We are removing it opportunisitically to
reduce the maintain burden.