Refactor specs to be adjacent to files
Why:
Currently we have our specs in a dedicated specs folder. Sometimes this
leads to confusion on which files have specs and which do not.
Additionally relative import paths in spec files can get somewhat long.
We would like to make an experiment by changing the folder structure for
specs such that specs are colocated with the file they test to determine
if this is a more ergonomic way to organize specs.