Add rules for three kings version
Why:
We would like to be able to play test the version of the game with the
rule of a player must take the opponents king three times to win.

This commit:
Updates the ThreeKingsContext to enforce this new rule. Additionally The
Piece interface definition was consolidated to be only in the
chess/types.ts file and not in both types.ts and pieces.ts.