Introduce useGameState hook
Why:
Currently we are shareing a log of stateful logic between our two game
providers. Since we are planning to add serveral more game modes, we
will want to be reusing this logic more. So we'd like to pull out common
game logic so that it may be reused.

Note: this is particially an experiment in custom hooks. It might turn
out that it's better to not abstract this logic out and simply duplicate
it.