Remove "0" from fen_id
Why:
Currently we have our FenId type including "0" to represent the empty
tile. This has led to some confusion and obfuscates the notion of the
empty tile not containing a a piece. A better design is to not include
"0" in the and instead allow tiles to optionally render a Piece or the
Empty tile.

This commit:
Makes the above described refactor and adds a new spec to the
ClassicContext to test that a human player is able to move their pieces.