Add try catch block to select tile function
Why:
There have been a few instances of crashes on test flight builds. We
don't have any error logging services setup currently and I am not able
to replicate the crashes locally. It has something to do with moving
pieces. We would like to protect from the crashes and log the crashes
somewhere.

This commit:
Adds a try catch block around the handelAttack call in selectTile in the
game_hooks logic. Hopefully this will save us from crashes until we can
figure out the source of the bug. Note that we are currently just
console.logging the error, which is somewhat pointless, we will need to
add error logging to an external service in a following commit. We are
simply adding the try catch block now, for sake of time.