Fix three kings bug
Why:
Currently it's possible to get into an unendable game state in the three
kings game mode, if you try to take your opponents king with your king.
Your king will be removed from the board, but it will still have lives,
so the game will never end.

This commit:
makes it so the game will consider a player with a missing king to have
lost rather than just a player with no lives. And makes it so that the
king will be removed if it has no lives left.