Allow users to change ai speed
Why:
While playing the game it would be more enjoyable for users if they are
able to experiment with which game speed they are interested in trying
out.

This commit:
Introduces the logic for changing the game speed for a game. A settings
modal was added with a slider component to accept user input in changing
the game speed. Note that the computerClockSpeed was moved into the
arcade context since that is a bit of state that we'd like to share
across games. (that is once you've selected a game speed, we'd like that
to be the game speed of your next game too.) Some logic was added so the
game pauses when the settings screen is opend. A GameContainer component
was added to handle the game screen components outside of the game
itself.