Add playground mode
Why:
We would like for players to have a stress-free zen version of the game.
We would also like a nice place for us to test out animation and
interaction ideas outside of the context of a game.

This commit:
Adds a playground mode to the game. adds a 'Go To Main Menu' button to
the settings modal so that users may end games early. Some refactoring
was done so that the countdown functionality could be included
independently of the game. The currently screen logic was refactored to
be more general, since we are planning to add more screens that Landing
and Game soon.