Add menu screen
Why:
We would like players to be able to select the game mode they'd like to
play from a dedicated menu screen.

This commit:
Introduces a menu screen for game modes. Updates the landing page to
have an 'Insert Coin' button to move to the menu screen. Updates the
game over screen to have an 'Insert Coin' Button and a return to the
main menu button.