Animate pieces on select
Why:
We would like for our game to have some animation when a piece is
selected so that the user experience is more engaging. We would like for
the piece to animate a bit larger when it is selected rather than the
tile border to being highlighted

this commit:
Introduces a basic animation to the piece component. We had to swap out
the existing svg assets with png assets to accomplish this since svgs
are not animatable by default in react-native. Note there are some
existing libraries that can handle svg animations, but they are
significantly more complicated that what is presented here so we are
deferening the work of animating the svgs to a later time.