Pull out tile component
Why:
We would like to add more animations to the pieces and tiles while the
user is playing the game.

This commit:
Pulls out a tile component from the board row component as a preliminary
refactor to make adding new functionality in a future commit easier. No
new functionality has been added.