Render pieces as siblings to tiles
Why:
We would like to start introducing animations for piece movement.
Currently the piece components are childern to the tile components which
makes this type of animation challenging because we lack control over
the pieces x and y position. In order to better manage the animation
logic it would be more convient to have the pieces be siblings to the
tiles and manually apply the x and y position of a piece using x and y
transition tranformations.

This commit:
Makes piece components sibling to tile components rather than children.
The tile width and height are cacluated based off of the screen
dimensions rather than using flex box so that we may know the ultimate
piece positions without needing to render the tiles first. That is, both
the tile components and piece components are based off of the screen and
not the react-native cacluated measure of the tile components.